fix(qbittorrent): decode HTML entities in torrent names
Some indexers emit `.torrent` files whose `name` field is HTML-escaped
(`Berl&iacute;n y la dama del armi&ntilde;o`), and the activity page
was rendering that verbatim — entities and all. Move the Latin-1 +
numeric-ref decoder out of TorznabService into a shared util and
apply it at the qBittorrent boundary too, so anything downstream
(queue activity, history matching, sourceTitle linking) receives the
human-readable form.
